fix: Output dbt STDOUT logs to Airflow task logging
For some reason dbt logs were only appearing in the worker logs, but
not in the task logs. This was causing the Airflow UI to not show any
dbt-specific logs, which was hiding important information.

This commit works around the issue by modifying dbt's log_manager to
stream STDOUT logs to a file, that is later read and logged using
the Airflow task's log.