feat: Full support for XCom by handling dbt's RunExecutionResult
Most dbt operations return a RunExecutionResult which is, as it is,
not JSON-serializable due to containing datetime.datetime objects. In
order to support moving run results via XCom, the DbtBaseOperator now
attempts to turn RunExecutionResult objects into JSON-serializable
dicts, by using dataclass.asdict and iso-formatting dates.

JSON is the default XCom backend, but if using custom XCom backends
the DbtBaseOpeartor.serializable_result may need to be overriden in a
subclass.