refactor: Split up DbtBaseOperator.execute
I was not keen on the amount of work DbtBaseOperator.execute was
doing: besides running the dbt command and handling the result, it was
also setting up logging and preparing the arguments. I've split up
these two operations into their own functions. This should make things
clearer while also provide a better ground to work on eventually doing
a proper override of dbt's logging.

This change also frees some space for me to work on S3 integration,
coming in the next commit.