feat: Enable pulling dbt projects and profiles from S3
This commit adds a new Airflow hook: DbtS3Hook. It subclasses from the
regular S3Hook, and implements a couple of methods:

* get_dbt_project: fetch all the contents from a dbt project directory
stored in S3.
* get_dbt_profiles: fetch a profiles.yml file stored in S3.

These methods will run whenever project_dir and/or profiles_dir is set
to an S3 URL. An optional s3_conn_id parameter was added to
DbtBaseOperator to specify which AWS connection to use.

The operators run the hook in a temporary directory, meaning that any
files downloaded will be cleaned up after Airflow is done, which helps
maintain task independence.