feat: Support for dbt 0.21 (#10)
* feat: Add support for new dbt build command and arguments

* feat: Support select argument for run, test, and compile operators

dbt 0.21 added the select flag to the run, test, and compile commands
with the intention for it to replace models and bring them in line
with the rest of the dbt commands. We have added support for a support
argument but kept backwards compatibility with models as we intend to
support dbt versions >= 0.19 (at least for the moment).