fix(s3): Skip empty files when downloading dbt projects from S3
Me and other users have noticed that when uploading files to S3 using
the AWS web UI an unnamed file may sometimes appear.

Not entirely sure if this is a bug on AWS' side or what, but it's
causing us issues when downloading dbt projects from S3 as we attemp
to create a file with the name of the parent directory (which usually
already exists, or is created and later downloads fail when they are
nested inside that file/directory).

So this deals with the issue by ignoring files like this. Also added a
unit test to cover this case, although it's kind of hard to test the
behaviour as this file apparently only appears when using the web UI.