feat(deps,docs): Allow checking for file changes before push to S3
DbtDepsOperator and DbtDocsGenerateOperator create artifacts that we
later push to S3. This commit adds optional safety checks that check
if a given list of files has been modified. This can be used by users
to ensure their projects have not been modified in unintended ways
before changes are persisted in S3.