feat(s3): Add parameter to control replacement of existing S3 keys
Setting replace_on_push to False will skip any S3 uploads when a key
already exists. This is useful to guarantee no unexpected files are
changed.

I'm still not super convinced about the solution as ideally
we would like partial replacing; only replacing changed files would be
ideal. This is hard to do with compressed projects, as it would
require downloading and uncompressing the project first. This has
been left for future work.