refactor: Make apache-airflow not required
Since we wish to support all MWAA supported versions of Airflow,
this requires supporting both 1.10.12 and 2.X releases. However, There
are several conflicting dependencies between dbt-core and
apache-airflow. None of the conflicts should break us, but they do
make dependency resolution impossible, which means airflow-dbt-python
cannot be installed.

For this reason we have removed the dependency to
apache-airflow. Assuming that as an Airflow operator we are always
installing airflow-dbt-python in an environment that already has
Airflow installed, this shouldn't cause any issues in production
deployments.

Moreover, this makes testing multiple Airflow versions easier as we
can pre-install apache-airflow.