fix(tests): Fix logging tests
Just realized that logging is still a big bugged: the dbt loggers are
producing multiple lines of output that are essentially the same, but
one of them is the debug output. This is caused by both the file
logger and the stdout logger being fired off. We tried to clear the
handlers from one of them in
operators.dbt.BaseDbtOperator.override_dbt_logging but this is not
working.

Due to this being fairly low priority, it will be fixed in a later commit.