chore(cicd): Alter testing strategy in GitHub Actions
Moving forward, the testing strategy will change. We will be targeting
all Python versions supported by dbt (>=3.7.2, <3.11 as of writing
this) and the following Airflow versions:

* 1.10.15 (latest 1.x release).
* 2.2.2 (latest MWAA supported version).
* Latest available version (2.4.* as of writing).

It's likely that support for 1.10.15 will be dropped soon. There isn't
much of a reason to be on Airflow 1 now that MWAA supports >2, which
was the original reason why we had support for 1.10.15.