feat: Partly implement new backend interface
This is the first in many commits in preparation for
airflow-dbt-python version 1.0.0.

The current objective is to move the temporary directory management
over from the base dbt operator to the base dbt hook. The idea behind
this change is that dbt in Airflow should always run in the context of
a temporary directory, so there is no reason to have this as an
operator depdendent implementation. It should be transparent to any
operator developer.

In line with this objective, I started with a few changes to the
backend interface to make it more generic with the introduction of a
URL class to identify the location of a dbt resource. This class
allows us to bridge the gap between backends that work with URLs and
backends that work with filesystems.

Moreover, dbt backends are now able to read files without storing them
first, although this will be utilized in a later commit.

Overall, the base backend class should now be doing work, with the
objective of simplifying the implementation of new backends.

A follow up commit will make backends proper hooks.