fix: Example DAG only tests models ran in the example itself
The example_complete_dbt_workflow was running all dbt tests, even
those that applied to models that were not run in the example (as they
are tagged as "deprecated").

This was causing tests to fail when running only the test module that
ran this example DAG. However, tests were passing when running the
entire test pipeline due to state from previous test modules

A follow-up commit will restrict the scope of the test database to the
test module, so as to avoid false positives like this one.