feat: Move temporary directory management to DbtHook
Managing a temporary directory to store dbt project files is an
intrinsic part of executing a dbt task: a task cannot -safely- run
without a temporary directory. This is because because we need:
* To keep tasks independent from the local filesystem.
* To keep task runs isolated from other task runs (including previous
instances of the same task).
* To ensure we clean up after we are done.

Due to temporary directories being inseparable from a dbt task
execution, operators shouldn't be concerned with a them; it should
managebe transparent for them.

This is specially true if we want to enable folks to build new
operators that do not subclass DbtBaseOperator, but simply use DbtHook
in their own workflows.

In conclusion, DbtHook is now in charge of creating a temporary directory.

There is more work to be done: follow up commits will revisit the way
we setup this directory, adding more work for DbtHook.