fix(cd): Properly fail deploy if CI fails
As part of our continuous deployment workflow, we have a step that
waits for our CI workflow to finish successfully. However, the action
used to wait for the CI workflow doesn't fail on a timeout, it just
sets an output variable.

So, we should have an extra step to check the output and fail
appropriately.

Finally, we shouldn't be timing out at all. CI takes roughly 20
minutes, so I'm bumping the timeout to an hour to make sure we have
plenty of time to wait for that.