fix: Use selector instead of selector_name
Around dbt version 1.5, the selector_name argument got renamed to
simply selector. Our configuration already supports it, so we should
use it everywhere and treat selector_name as a backwards compatible
alias.

Moreover, selector is always a string, and not a list of strings as it
was typed in the DbtRunOperator.