fix: Pass empty list as 'active_integrations' in 'parse_manifest'
This should allow code running dbt 1.10.7 or later not to break.

We still haven't added actual support for whatever this argument is
doing, this is just ensuring we don't crash by setting the necessary
argument. Some investigation is required into the new feature/argument
to figure out what to pass to it.